refactor: Domain.Core → Domain.Services
