--- Assets/YukimaruGames/Terminal/Runtime/Domain/Core/Commands.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9b48bd77acd93524eaa7e3d4bca98255
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
