+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Services/CommandAutocomplete.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using YukimaruGames.Terminal.Domain.API.Commands;
 
-namespace YukimaruGames.Terminal.Domain.Core.Commands
+namespace YukimaruGames.Terminal.Domain.Services
 {
     /// <summary>
     /// コマンドの自動補完クラス.
