+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Services/CommandInvoker.cs
@@ -1,6 +1,6 @@
 using YukimaruGames.Terminal.Domain.API.Commands;
 
-namespace YukimaruGames.Terminal.Domain.Core.Commands
+namespace YukimaruGames.Terminal.Domain.Services
 {
     /// <summary>
     /// コマンドの実行クラス.
