+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Services/YukimaruGames.Terminal.Domain.Services.asmdef
@@ -1,6 +1,6 @@
 {
-    "name": "YukimaruGames.Terminal.Domain.Core",
-    "rootNamespace": "YukimaruGames.Terminal.Domain.Core",
+    "name": "YukimaruGames.Terminal.Domain.Services",
+    "rootNamespace": "YukimaruGames.Terminal.Domain.Services",
     "references": [
         "GUID:37e9bf61fa8119c4c8e79d1f08c439aa",
         "GUID:b91bfe61ad34c074d9f0d4e53cb288f0"
