+++ Assets/YukimaruGames/Terminal/Runtime/Runtime/Installer/TerminalNullInstaller.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Linq;
 using YukimaruGames.Terminal.Application.Core;
-using YukimaruGames.Terminal.Domain.Core.Commands;
+using YukimaruGames.Terminal.Domain.Services;
 using YukimaruGames.Terminal.Runtime.Shared;
 using YukimaruGames.Terminal.SharedKernel;
